requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
websockets==11.0.3
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4
python-dotenv==1.0.0
pydantic==2.4.2
openai==1.3.0
edge-tts==6.1.9
aiofiles==23.2.1
numpy==1.26.0
requests==2.31.0
pydub==0.25.1